% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{as_change_label}
\alias{as_change_label}
\title{Calculate change labels of lagged change scores}
\usage{
as_change_label(
  value,
  lag_value,
  change_label = c("high_low", "deterio_improve")
)
}
\arguments{
\item{lag_value}{}
}
\value{

}
\description{
Calculate change labels of lagged change scores
}
\examples{
as_change_label(value = 0, lag_value = 4,
                change_label = "deterio_improve")
}
