% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptions.R
\name{as_i8_desc}
\alias{as_i8_desc}
\title{Convert HoNOS item 8 types to labelled factor}
\usage{
as_i8_desc(x, short_descriptions = TRUE)
}
\arguments{
\item{x}{Vector of strings}

\item{short_descriptions}{Logical, specifying whether to return short or long descriptions}
}
\value{
Labelled "factor"
}
\description{
Convert HoNOS item 8 types to labelled factor
}
\examples{
x <- c("A", "B", "C", "D", "E", "F", "G", "H", "I", "J")
as_i8_desc(x)
}
