% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptions.R
\name{as_item_desc}
\alias{as_item_desc}
\title{Convert HoNOS item numbers to labelled factor}
\usage{
as_item_desc(x, n_items)
}
\arguments{
\item{x}{Vector of item numbers}

\item{n_items}{Numeric, specifying maximum number of items}
}
\value{

}
\description{
Convert HoNOS item numbers to labelled factor
}
\examples{
x <- 1:13
as_item_desc(x)
}
