% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptions.R
\name{as_severity_desc}
\alias{as_severity_desc}
\title{Convert HoNOS severity scores to labelled factor}
\usage{
as_severity_desc(x, short_labels = TRUE, na_level = NULL)
}
\arguments{
\item{x}{Vector of severity rating from 0 to 4}

\item{short_labels}{Logical, specifying whether to return short or long descriptions}

\item{na_level}{String, specifying label for missing values.}
}
\value{
Labelled "factor"
}
\description{
Convert HoNOS severity scores to labelled factor
}
\examples{
x <- 0:4
as_severity_desc(x)
}
