% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calc_change_label}
\alias{calc_change_label}
\title{Calculate change labels for lagged change scores}
\usage{
calc_change_label(
  value,
  lag_value,
  change_label = c("high_low", "deterio_improve")
)
}
\arguments{
\item{value}{Name of variable with HoNOS score at time point (t)}

\item{lag_value}{Name of variable with HoNOS score at previous time point (t-1)}

\item{change_label}{String, specifying whether to describe the change using "high_low" (e.g., HL stands for high to low) or "deterio_improve" (Deterioration, Unchanged, Improved)}
}
\value{
Labelled "factor"
}
\description{
Calculate change labels for lagged change scores
}
\examples{
calc_change_label(value = 0, lag_value = 4,
                change_label = "deterio_improve")
}
