% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calc_subscales}
\alias{calc_subscales}
\title{Calculate sum scores of HoNOS subscales (12-item version)}
\usage{
calc_subscales(
  data,
  id_var,
  date_var,
  item_var,
  value_var,
  return_format = c("long", "wide"),
  return_items = FALSE
)
}
\arguments{
\item{data}{Data in long format}

\item{id_var}{Name of variable that uniquely identifies each individual}

\item{date_var}{Name of date (or datetime) variable}

\item{item_var}{Name of variable specifying item numbers}

\item{value_var}{Name of variable with HoNOS scores}

\item{return_format}{String, specifying whether to return data in long (i.e., TODO) or wide (i.e., TODO) format}

\item{return_items}{Logical, specifying whether to include individual item scores. NOTE: CURRENTLY NOT SUPPORTED.}
}
\value{

}
\description{
Calculate sum scores of HoNOS subscales (12-item version)
}
