% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{lag_honos}
\alias{lag_honos}
\title{Calculate lagged HoNOS scores}
\usage{
lag_honos(
  data,
  id_var,
  date_var,
  add_change_label = TRUE,
  change_label = c("high_low", "deterio_improve")
)
}
\arguments{
\item{change_label}{}
}
\value{

}
\description{
Calculate lagged HoNOS scores
}
\examples{
lag_honos(data = honos_wide,
          id_var = id,
          date_var = date,
          change_label = "deterio_improve")
}
