% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{rename_honos}
\alias{rename_honos}
\title{Helper function to rename HoNOS variables}
\usage{
rename_honos(
  value_vars_current,
  prob_var_item8,
  spec_var_item8,
  value_vars_history,
  honos_version = c("working_adults")
)
}
\arguments{
\item{value_vars_current}{Vector, specifying variable names with values for 'current' items}

\item{prob_var_item8}{Vector, specifying variable name with description of problem (prob) for item 8}

\item{spec_var_item8}{Vector, specifying variable name with problem specification (spec) of for item 8}

\item{value_vars_history}{Vector, specifying specifying variable names with values for 'historic' items}

\item{honos_version}{Vector, specifying version of the HoNOS that is being used. TODO IMPLEMENT MORE VERSIONS.}
}
\value{

}
\description{
Helper function to rename HoNOS variables
}
