% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{lag_honos}
\alias{lag_honos}
\title{Calculate lagged HoNOS scores}
\usage{
lag_honos(
  data,
  id_var,
  date_var,
  add_change_label = TRUE,
  change_label = c("high_low", "deterio_improve")
)
}
\arguments{
\item{data}{Dataframe in wide format (i.e., one row for each HoNOS assessment and one column for each HoNOS item)}

\item{id_var}{Name of variable that uniquely identifies each individual}

\item{date_var}{Name of date (or datetime) variable}

\item{add_change_label}{Logical, specifying whether to add a variable that describes the change using the function \code{\link{calc_change_label}}}

\item{change_label}{String, specifying whether to describe the change using "high_low" (e.g., HL stands for high to low) or "deterio_improve" (Deterioration, Unchanged, Improved)}
}
\value{
Dataframe with original data and lagged values
}
\description{
Calculate lagged HoNOS scores
}
\examples{
lag_honos(data = honos_wide,
          id_var = id,
          date_var = date,
          change_label = "deterio_improve")
}
